Code cleanup and trying to find 01:00 bug
"01:00" bug stops the whole MCU after printing "01:00" to the
watchface.

But this only happens when starting at 00:59. Starting at 01:00
works fine. 01:59 also works fine.

+++ firmware/Src/clock.c
@@ -7,7 +7,7 @@ extern RTC_HandleTypeDef hrtc;
 bool needTimeUpdate(applicationState_t* state) {
   RTC_TimeTypeDef sTime;
   bool ret = false;
-  uint32_t nextTick = 0;
+  uint8_t nextTick = 0;
 
   if (HAL_RTC_GetTime(&hrtc, &sTime, RTC_FORMAT_BIN) != HAL_OK) {
     Error_Handler();
